refactor styles: move CSS files to a dedicated styles directory and remove unused styles
